--- src/components/Shared/IconText/index.js
@@ -1,1 +1,0 @@
-export { default } from "./IconText";
