--- src/components/Shared/ProgressBar/index.js
@@ -1,1 +1,0 @@
-export { default } from "./ProgressBar";
