--- src/components/Shared/TabPanel/TabPanel.js
@@ -1,25 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-const TabPanel = ({ children, value, index, ...props }) => {
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      className={`${props.className}`}
-      {...props}
-    >
-      {value === index && children}
-    </div>
-  );
-};
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.any.isRequired,
-  value: PropTypes.any.isRequired
-};
-
-export default TabPanel;
