--- src/components/Shared/TabPanel/index.js
@@ -1,1 +1,0 @@
-export { default } from "./TabPanel";
