--- src/components/package/PackageDependencies/index.js
@@ -1,1 +1,0 @@
-export { default } from "./PackageDependencies";
