--- src/components/package/PackageHeader/index.js
@@ -1,1 +1,0 @@
-export { default } from "./PackageHeader";
