--- src/components/package/PackageLicenses/index.js
@@ -1,1 +1,0 @@
-export { default } from "./PackageLicenses";
