+++ src/features/advise/routes/AdviseNotFound.js
@@ -1,0 +1,3 @@
+export const AdviseNotFound = ({analysis_id}) => {
+    return <div>"{analysis_id}" not found</div>;
+};
