+++ src/features/misc/api/index.js
@@ -1,1 +1,2 @@
 export * from "./getPackageMetadata"
+export * from "./getFile"
