+++ src/features/package/components/PackageDependencies/index.js
@@ -1,0 +1,1 @@
+export * from "./PackageDependencies";
