+++ src/features/package/components/PackageHeader/index.js
@@ -1,0 +1,1 @@
+export * from "./PackageHeader";
