+++ src/features/package/components/PackageLicenses/PackageLicenses.md
