+++ src/features/package/components/PackageLicenses/index.js
@@ -1,0 +1,1 @@
+export * from "./PackageLicenses";
