+++ src/features/package/components/index.js
@@ -1,0 +1,3 @@
+export * from "./PackageHeader"
+export * from "./PackageDependencies"
+export * from "./PackageLicenses"
