+++ src/features/package/routes/PackageNotFound.js
@@ -1,0 +1,3 @@
+export const PackageNotFound = ({package_name, package_version}) => {
+    return <div>"{package_name}{package_version ? " " + package_version : ""}" not found</div>;
+};
