--- src/navigation/CONSTANTS.js
@@ -1,4 +1,0 @@
-export const REPONAME = "/thoth-search";
-export const ROOT = "/";
-export const ADVISE = "/advise";
-export const PACKAGE = "/package";
