--- src/navigation/NotFound.js
@@ -1,7 +1,0 @@
-import { Typography } from "@material-ui/core";
-
-export const NotFound = ({ text }) => {
-  return <Typography variant="h2">404: {text ?? "page"} not found!</Typography>;
-};
-
-export default NotFound;
