--- src/pages/Dashboard/hooks/index.js
@@ -1,4 +1,0 @@
-export { useLockFileToGraph } from "./useLockFileToGraph";
-export { useMergeGraphs } from "./useMergeGraphs";
-export { useComputeMetrics } from "./useComputeMetrics";
-export { useInterval } from "./useInterval";
