+++ src/routes/RouterConfig.js
@@ -2,19 +2,16 @@ import {BrowserRouter, Route, Routes} from "react-router-dom";
 import {Home} from "features/home";
 import {NotFound} from "features/misc";
 import {AdviseRoutes} from "features/advise";
+import {PackageRoutes} from "features/package";
 
 export const RouterConfig = () => {
   return (
       <BrowserRouter basename={"/thoth-search"}>
           <Routes>
             <Route exact path={"/"} element={<Home />} />
-            <Route path={"advise/:analysis_id/*"} element={<AdviseRoutes />} />
-
-            {/*<Route*/}
-            {/*exact*/}
-            {/*path={"package/:package_name/:package_version?"}*/}
-            {/*component={PackageOverview}*/}
-            {/*/>*/}
+            <Route exact path={"advise/:analysis_id/*"} element={<AdviseRoutes />} />
+            <Route path={"package/:package_name/*"} element={<PackageRoutes />}
+            />
             <Route path="*" status={404} element={<NotFound/>} />
           </Routes>
       </BrowserRouter>
