--- src/services/CONSTANTS.js
@@ -1,3 +1,0 @@
-export const PYPI = "https://pypi.org/pypi";
-export const THOTH = "https://khemenu.thoth-station.ninja/api/v1";
-export const LICENSES = "https://raw.githubusercontent.com/spdx/license-list-data/master/json/licenses.json";
